No looping path for repeat with at most one repeat
Repeat used to display a looping path under the block regardless of the maximum number of times it can be used. When this maximum is one like with `c?` this looping path is actually useless.

Add a condition for the maximum and do not display the looping path is maximum is one. If minimum is 0, it already display a short circuit path on top of the block so both `{0, 1}` and `{1, 1}` will display properly.